% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGOdoubleBarGraph.R
\name{DGObarplot}
\alias{DGObarplot}
\title{Draw a bar graph on the DO and GO analysis.}
\usage{
DGObarplot(
  DGOResult,
  showCategory = 8,
  DOcol = c("red", "darksalmon"),
  GOcol = c("blue", "lightskyblue"),
  pAdjustCutoff = 0.05
)
}
\arguments{
\item{DGOResult}{DO and GO enrichment analysis result returned from enrichDGO().}

\item{showCategory}{number of ontology groups to show from DO and GO each.}

\item{DOcol}{color to use for plotting DO.}

\item{GOcol}{color to use for plotting GO.}

\item{pAdjustCutoff}{p.adjust cutoff value for ontology group to plot.}
}
\value{
Returns a double bar graph of DO and GO analysis.
\itemize{
\item different color for DO and GO results
\item color gradient by p.adjust value
\item ordered by p.adjust value
}
}
\description{
A function that draws a double bar graph
of DO and GO analysis, ordered and colored by group and p.adjust value.
}
\examples{
\dontrun{
# load data from DOSE
library(DOSE)
data(geneList)
gene <- names(geneList)[abs(geneList) > 2]
result <- enrichDGO(gene, universe=names(geneList))

# tamper the data little bit to see 
# the full functionalities of DGOplot
result$DO@result$p.adjust <-  result$DO@result$p.adjust / 10
 
DGObarplot(result)
DGObarplot(result,
           DOcol = c("purple", "yellow"),
           GOcol = c("red", "blue"),
           showCategory = 6)
}

}
