% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGOassociationNetwork.R
\name{DGOnetplot}
\alias{DGOnetplot}
\title{Draw a gene association network of DO and GO enrichment analyses results.}
\usage{
DGOnetplot(DGOResult, showCategory = 6, pvalueCutoff = 0.05)
}
\arguments{
\item{DGOResult}{DO and GO enrichment analysis result returned from enrichDGO().}

\item{showCategory}{number of ontology groups to show from DO and GO each.}

\item{pvalueCutoff}{p-value cutoff.}
}
\value{
Returns a gene association network of DO and GO analyses.
\itemize{
\item edges colored in gradient according to p.adjust value
\item edges to DO and GO terms colored differently
\item ontology term nodes size proportional to number of genes associated with it
}
}
\description{
A function that draws a gene association network with DO and GO terms from enrichment analyses.
}
