% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGOEnrichmentAnalysis.R
\name{enrichDGO}
\alias{enrichDGO}
\title{Perform DO and GO enrichment analysis}
\usage{
enrichDGO(gene, Gont = "MF", Dont = "DO", OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe,
  qvalueCutoff = 0.2, minGSSize = 10, maxGSSize = 500,
  readable = FALSE, pool = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{Gont}{GO categories; one of "BP", "MF", "CC", or "ALL" for all three.}

\item{Dont}{DO categories; one of DO or DOLite}

\item{OrgDb}{OrgDb}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{qvalueCutoff}{q-value cutoff}

\item{minGSSize}{minimal size of genes annotated by NCG category for testing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{readable}{whether mapping gene ID to gene Name}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies}

\item{pValueCutoff}{p-valuve cutoff}
}
\value{
Returns a list of 2 enrichResults
\itemize{
\item DOanalysis - an enrichResult object with DO enrichment analysis result
\item GOanalysis - an enrichResult object with DO enrichment analysis result
}
}
\description{
A function that performs DO and GO enrichment analysis given a list of geneIDs.
}
\examples{
data(geneList)  # load data from DOSE package
gene <- names(geneList)[abs(geneList) > 2]
DGOResult <- enrichDGO(gene, universe = names(geneList))
DOanalysis <- DGOResult["DO"]
GOanalysis <- DGOResult["GO"]

}
