% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{makeLegend}
\alias{makeLegend}
\title{MAkes a legend for DGOnetplot}
\usage{
makeLegend(descr)
}
\arguments{
\item{descr}{a vector of gene ontology names}
}
\value{
Returns a legend to use for DGPnetplot
}
\description{
A helper function that parses the ontology names to generate a legend.
}
