% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphLayout.R
\name{netLayout}
\alias{netLayout}
\title{A simplified function from NetPathMiner by ahmohamed reconstructed
https://stackoverflow.com/questions/28715736/how-to-spread-out-community-graph-made-by-using-igraph-package-in-r/28722680#28722680}
\usage{
netLayout(graph, wc, cluster.strength = 10, layout = layout.auto)
}
\arguments{
\item{graph}{garph}

\item{wc}{number of vertices added for algorithm}

\item{cluster.strength}{cluster strength}

\item{layout}{layout}
}
\value{
Returns a layout for the gene association network
}
\description{
A helper function with an algorithm to spread out the nodes in an association network.
}
