% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDGOplot.R
\name{runDGOplot}
\alias{runDGOplot}
\title{Launch the shiny app for package DGOplot}
\usage{
runDGOplot()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runTestingPackage()
}

}
