% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WarningShortEnrichResult.R
\name{warnOntN}
\alias{warnOntN}
\title{Warning or Error generation for the number of ontology groups input.}
\usage{
warnOntN(nDO, nGO, showCategory)
}
\arguments{
\item{nDO}{number of DO enrichment result groups}

\item{nGO}{number of GO enrichment result groups}

\item{showCategory}{number of groups to show on the plot}

\item{type}{"bar" for DGObarplot and "net" for DGOnetplot}
}
\value{
warning or error depending on the input
}
\description{
A helper function that checks the number of ontology groups
and if showCategory exceeds them.
}
